RFC combine matches and progress
This is a non-functional RFC commit to discuss a change to the streaming
API. I realised a problem of the current protocol is that progress and
matches events can come in at different times. This can lead to updating
the list of results shown without updating progress. Or vice versa. To
avoid that we group both results and matches together.

We can go more extreme and include alert and filters in the results
event. This may simplify stuff. WDYT?